\name{buster-package}
\alias{buster-package}
\docType{package}
\title{
Performs bagging on hierarchical clusters
}
\description{
More about what it does (maybe more than one line)
z
}
\details{
\tabular{ll}{
Package: \tab buster\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-03\cr
License: \tab What license is it under?\cr
}
y
}
\author{
Simon Raper
}
\references{
x
}

